Disable end game condition; reset overclocks at end of round
Game was crashing because after buying a card/tile, end condition was met and EndGame was throwing NotImplementedException.

Current overclock implementation removes them from player's board (as far as I can tell). This will need to be changed to use overclock function Use() and boolean HasBeenUsed to determine which overclocks are available.